+++ Tilli/Model/SalesAnalyticsModel.swift
@@ -13,16 +13,16 @@ struct HourlyAnalysisData: Identifiable {
     let id = UUID()
     let hour: Int  // 0-23
     let hourString: String  // "09:00"
-    let amount: Double  // 該時段總金額
+    let amount: Decimal  // 該時段總金額
     let transactions: Int  // 該時段交易筆數
-    let avgPrice: Double  // 平均客單價
+    let avgPrice: Decimal  // 平均客單價
 
-    init(hour: Int, amount: Double, transactions: Int) {
+    init(hour: Int, amount: Decimal, transactions: Int) {
         self.hour = hour
         self.hourString = String(format: "%02d:00", hour)
         self.amount = amount
         self.transactions = transactions
-        self.avgPrice = transactions > 0 ? amount / Double(transactions) : 0
+        self.avgPrice = transactions > 0 ? amount / Decimal(transactions) : 0
     }
 }
 
@@ -31,11 +31,11 @@ struct PaymentMethodAnalysisData: Identifiable {
     let method: PaymentMethod
     let name: String
     let transactions: Int
-    let amount: Double
+    let amount: Decimal
     let percentage: Int
     let color: Color
 
-    init(method: PaymentMethod, transactions: Int, amount: Double, totalTransactions: Int) {
+    init(method: PaymentMethod, transactions: Int, amount: Decimal, totalTransactions: Int) {
         self.method = method
         self.transactions = transactions
         self.amount = amount
@@ -53,17 +53,17 @@ struct PaymentMethodAnalysisData: Identifiable {
 }
 
 struct SalesOverviewData {
-    let totalAmount: Double
+    let totalAmount: Decimal
     let totalTransactions: Int
-    let avgTransactionValue: Double
+    let avgTransactionValue: Decimal
     let peakHour: Int
-    let peakHourAmount: Double
+    let peakHourAmount: Decimal
     let paymentMethodStats: [PaymentMethodAnalysisData]
 
-    init(totalAmount: Double, totalTransactions: Int, peakHour: Int, peakHourAmount: Double, paymentMethodStats: [PaymentMethodAnalysisData]) {
+    init(totalAmount: Decimal, totalTransactions: Int, peakHour: Int, peakHourAmount: Decimal, paymentMethodStats: [PaymentMethodAnalysisData]) {
         self.totalAmount = totalAmount
         self.totalTransactions = totalTransactions
-        self.avgTransactionValue = totalTransactions > 0 ? totalAmount / Double(totalTransactions) : 0
+        self.avgTransactionValue = totalTransactions > 0 ? totalAmount / Decimal(totalTransactions) : 0
         self.peakHour = peakHour
         self.peakHourAmount = peakHourAmount
         self.paymentMethodStats = paymentMethodStats
